// src/model/surrounding/ServerConfig.ts
export interface ServerConfig {
    createdBy: string;
    createdDate: Date;
    updatedBy: string;
    updatedDate: Date;
    idConfigMain: string;
    keyGroup: string;
    keyName: string;
    value: number;
    description: string;
    statusCode: string;
    typeForm: string;
    sortField: number;
}