// src/model/surrounding/ConfigParameter.ts
export interface ConfigParameter {
    createdBy: string;
    createdDate: Date;
    updatedBy: string;
    updatedDate: Date;
    idParameter: string;
    groupCode: string;
    paramCode: string;
    paramName: string;
    paramDesc: string;
    sortNumber: number;
    active: boolean;
    typeForm: string;
}